feat: pathfinder refactor
